refactor: Player에게 계산 결과 책임 위임
