+++ src/test/java/domain/BetTest.java
@@ -1,7 +1,5 @@
-package domain.card;
+package domain;
 
-import domain.Bet;
-import domain.MatchResult;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
